(edit) Breakup the homepage and create HomeDocument/HomeExamples/HomeFeatures/HomeHero/HomeSignUp components
